Makefile, libft.h, ft_strlcpy

+++ ft_bzero.c
@@ -6,15 +6,15 @@
 /*   By: leolipes <leolipes@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/07/29 14:10:16 by leolipes          #+#    #+#             */
-/*   Updated: 2021/07/29 14:10:17 by leolipes         ###   ########.fr       */
+/*   Updated: 2021/08/03 13:01:10 by leolipes         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <unistd.h>
+#include "libft.h"
 
 void	ft_bzero(void *str, size_t n)
 {
-	int		i;
+	size_t		i;
 	char	*ptr;
 
 	i = 0;
