+++ ft_memcpy.c
@@ -6,15 +6,15 @@
 /*   By: leolipes <leolipes@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/07/29 14:14:04 by leolipes          #+#    #+#             */
-/*   Updated: 2021/07/29 14:52:07 by leolipes         ###   ########.fr       */
+/*   Updated: 2021/08/03 13:03:06 by leolipes         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <unistd.h>
+#include "libft.h"
 
 void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n)
